requirements.txt
alembic==1.13.1
aniso8601==9.0.1
atomicwrites==1.4.1
attrs==23.2.0
backcall==0.2.0
blinker==1.8.2
click==8.1.7
coverage==7.5.3
decorator==5.1.1
Faker==0.8.15
flake8==3.5.0
Flask==3.0.3
flask-marshmallow==1.2.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.10
flask-shell-ipython==0.3.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==19.7.1
iniconfig==2.0.0
ipython==6.4.0
isort==4.3.4
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
mccabe==0.6.1
more-itertools==10.2.0
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pluggy==1.5.0
prompt-toolkit==1.0.18
psycopg2-binary==2.9.9
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.18.0
pytest==8.2.1
pytest-cov==2.5.1
pytest-flask==1.3.0
pytest-runner==4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
setuptools==70.0.0
simplegeneric==0.8.1
six==1.16.0
SQLAlchemy==2.0.30
text-unidecode==1.2
traitlets==5.14.3
typing_extensions==4.12.0
wcwidth==0.2.13
Werkzeug==3.0.3
WTForms==3.1.2
